Update Tracker.py
I took the one loop that I had before and changed it into 2 separate loops for different positions in the program. One is before the button has been pressed and the other is after the button has been pressed once, and is waiting to pressed again to end the program. Also I added loops to have the names of the files be different so they don't override old files with the same name.